feat(vpc): add initial VPC management and improve session performance

+++ Core/session.sh
@@ -1,3 +1,5 @@
+CURRENT_USER="Guest"
+ACCOUNT_ID="N/A"
 login() {
 
     echo ""
@@ -20,18 +22,79 @@ EOF
 region=$AWS_REGION
 output=json
 EOF
+IDENTITY=$(
+    aws sts get-caller-identity 2>/dev/null
+)
 
-    if aws sts get-caller-identity >/dev/null 2>&1
-    then
-        echo ""
-        echo "[OK] Login Successful"
+if [ $? -eq 0 ]; then
+
+    ACCOUNT_ID=$(echo "$IDENTITY" | jq -r '.Account')
+    ARN=$(echo "$IDENTITY" | jq -r '.Arn')
+
+    if [[ "$ARN" == *":root" ]]; then
+        CURRENT_USER="Root"
     else
-        echo ""
-        echo "[ERROR] Invalid Credentials"
+        CURRENT_USER=$(basename "$ARN")
+    fi
+
+    echo ""
+    echo "[OK] Login Successful"
+
+else
+
+    CURRENT_USER="Guest"
+    ACCOUNT_ID="N/A"
+
+    echo ""
+    echo "[ERROR] Invalid Credentials"
+
+fi
+
+}
+load_session() {
+
+    IDENTITY=$(
+        aws sts get-caller-identity 2>/dev/null
+    )
+
+    if [ $? -ne 0 ]; then
+        CURRENT_USER="Guest"
+        ACCOUNT_ID="N/A"
+        return
+    fi
+
+    ACCOUNT_ID=$(echo "$IDENTITY" | jq -r '.Account')
+    ARN=$(echo "$IDENTITY" | jq -r '.Arn')
+
+    if [[ "$ARN" == *":root" ]]; then
+        CURRENT_USER="Root"
+    else
+        CURRENT_USER=$(basename "$ARN")
     fi
 }
 
 logout() {
+
+    if [ "$CURRENT_USER" = "Guest" ]; then
+        echo ""
+        echo "[INFO] No active session."
+        return
+    fi
+
+    echo ""
+    echo "================================="
+    echo "          LOGOUT"
+    echo "================================="
+    echo ""
+    echo "Current User : $CURRENT_USER"
+    echo ""
+    read -r -p "Type LOGOUT to continue: " CONFIRM
+
+    if [ "$CONFIRM" != "LOGOUT" ]; then
+        echo ""
+        echo "[INFO] Logout cancelled."
+        return
+    fi
 
     rm -f ~/.aws/credentials
 
@@ -41,26 +104,14 @@ region=$AWS_REGION
 output=json
 EOF
 
+    CURRENT_USER="Guest"
+    ACCOUNT_ID="N/A"
+
     echo ""
-    echo "[OK] Logged Out"
+    echo "[OK] Logged out successfully."
 }
-
 current_identity() {
-
-    ARN=$(aws sts get-caller-identity \
-        --query Arn \
-        --output text 2>/dev/null)
-
-    if [ $? -ne 0 ]; then
-        echo "Guest"
-        return
-    fi
-
-    if [[ "$ARN" == *":root" ]]; then
-        echo "Root"
-    else
-        basename "$ARN"
-    fi
+ echo "$CURRENT_USER"
 }
 
 check_session() {
